+++ tests/sensors/test_sensors.py
@@ -36,7 +36,7 @@ MOCK_FIVETRAN_RESPONSE_PAYLOAD = {
         "connected_by": "mournful_shalt",
         "created_at": "2021-03-05T22:58:56.238875Z",
         "succeeded_at": "2021-03-23T20:55:12.670390Z",
-        "failed_at": 'null',
+        "failed_at": "null",
         "sync_frequency": 360,
         "schedule_type": "manual",
         "status": {
@@ -45,7 +45,7 @@ MOCK_FIVETRAN_RESPONSE_PAYLOAD = {
             "update_state": "on_schedule",
             "is_historical_sync": False,
             "tasks": [],
-            "warnings": []
+            "warnings": [],
         },
         "config": {
             "latest_version": "1",
@@ -53,34 +53,34 @@ MOCK_FIVETRAN_RESPONSE_PAYLOAD = {
             "named_range": "fivetran_test_range",
             "authorization_method": "User OAuth",
             "service_version": "1",
-            "last_synced_changes__utc_": "2021-03-23 20:54"
-        }
-    }
+            "last_synced_changes__utc_": "2021-03-23 20:54",
+        },
+    },
 }
 
 
 # Mock the `conn_fivetran` Airflow connection (note the `@` after `API_SECRET`)
-@mock.patch.dict('os.environ', AIRFLOW_CONN_CONN_FIVETRAN='http://API_KEY:API_SECRET@')
+@mock.patch.dict("os.environ", AIRFLOW_CONN_CONN_FIVETRAN="http://API_KEY:API_SECRET@")
 class TestFivetranSensor(unittest.TestCase):
-    """ 
-    Test functions for Fivetran Operator. 
+    """
+    Test functions for Fivetran Operator.
 
     Mocks responses from Fivetran API.
     """
 
-    @mock.patch.object(FivetranSensor, 'poke', 'returned_sync_status')
+    @mock.patch.object(FivetranSensor, "poke", "returned_sync_status")
     @requests_mock.mock()
     def test_del(self, m):
 
         sensor = FivetranSensor(
-            task_id='my_fivetran_sensor',
-            fivetran_conn_id='conn_fivetran',
-            connector_id='interchangeable_revenge'
+            task_id="my_fivetran_sensor",
+            fivetran_conn_id="conn_fivetran",
+            connector_id="interchangeable_revenge",
         )
 
         log.info(sensor.poke)
-        assert sensor.poke == 'returned_sync_status'
+        assert sensor.poke == "returned_sync_status"
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     unittest.main()